pyproject.toml
[tool.poetry]
name = "cookiecutter-cruft-poetry-tox-pre-commit-ci-cd"
version = "0.6.0"
description = "Standardized Python project creation and synchronization"
authors = ["Teo Zosa <erinzosa@ucla.edu>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/TeoZosa/cookiecutter-cruft-poetry-tox-pre-commit-ci-cd"
homepage = "https://github.com/TeoZosa/cookiecutter-cruft-poetry-tox-pre-commit-ci-cd"
keywords = [
    "Cookiecutter",
    "Cruft",
    "Dependabot",
    "GitHub Actions",
    "Poetry",
    "Pre-commit",
    "Python",
    "Release Drafter",
    "Tox"]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python",
]
# Dummy package for `pip install .` compatibility
packages = [
    { include = "{{cookiecutter.package_name}}", from = "{{cookiecutter.project_slug}}" },
]

[tool.poetry.dependencies]
python = "^3.8,<3.12"
# Project boilerplate synchronizer
cruft = "^2.11.1"

# Project cookiecutter dependencies
hypothesis = "^6.61.0"
python-dotenv = "^0.21.0"
importlib-metadata = "^5.1.0"
structlog-sentry-logger = "^0.18.0"
typer = {extras = ["all"], version = "^0.6.1"}

[tool.poetry.group.dev.dependencies]
# Type Checking and Data Validation
mypy = "^0.991" # Static type checker

# Testing
pytest = "^7.2.0"
pytest-benchmark = "^3.4.1"
pytest-cov = "^4.0.0"
pytest-emoji = "^0.2.0"
pytest-mock = "^3.10.0"
pytest-sugar = "^0.9.6"
pytest-xdist = "^3.0.2"
xdoctest = "^1.1.0"
tox = "^3.27.1"

# Linting
## Code formatting
black = "^22.10.0" # see: https://black.readthedocs.io/en/stable/editor_integration.html
## Code quality
pylint = "^2.15.5"
## Automation and management
pre-commit = "^2.20.0"

# Documentation
darglint = "^1.8.1"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
emoji = "^1.7.0"
myst-parser = "^0.18.0"
pygments = "^2.13.0"
sphinx = "^5.1.1"
sphinx-autodoc-typehints = "^1.19.1"
sphinx-rtd-theme = "^1.0.0"
sphinxcontrib-apidoc = "^0.3.0"
types-emoji = "^1.7.1"

#################################################################################
# Tooling configs                                                                #
#################################################################################
[tool.black]
line-length = 120

[tool.flakeheaven]
# specify any flake8 options
exclude = [".*"] # Ignore dotfiles
# make output nice
format = "grouped"
# show line of source code in output
show_source = true

# list of plugins and rules for them
[tool.flakeheaven.plugins]
# cyclomatic complexity (https://github.com/PyCQA/mccabe)
mccabe = ["+*"]
# include everything in pyflakes except F401
pyflakes = ["+*", "-F401"]
# include everything in pycodestyle except what Black covers
pycodestyle = ["+*",
    "-E203", # Whitespace before ":"
    "-E501", # Line too long (82 > 78 characters)
    "-W503"  # Line break occurred before a binary operator <- this is now considered best practice by PEP 8
]

[tool.isort]
profile = "black"
atomic = true
combine_as_imports = true

[mypy]
ignore_missing_imports = true
pretty = true
show_column_numbers = true
show_error_context = true
show_error_codes = true

[tool.pylint.basic]
good-names-rgxs = ["^Test_.*$"]

[tool.pylint.messages_control]
disable = [
  # Explicitly document only as needed
  "missing-module-docstring",
  "missing-class-docstring",
  "missing-function-docstring",
  # Black & Flake8 purview
  "line-too-long",
  "c-extension-no-member",
]

[tool.pylint.similarities]
# Ignore imports when computing similarities.
ignore-imports = "yes"

[tool.pytest.ini_options]
addopts = ["-rfsxX", "-l", "--tb=short", "--strict-markers", "-vv", "--emoji", "--xdoctest"]
xfail_strict = true
testpaths = ["tests",]
norecursedirs = [".*", "*.egg", "build", "dist",]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
